+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # MatchZy Changelog
 
+# 0.8.14
+
+#### October 21, 2025
+
+- Bumped CSS Version.
+- Added .savepos and .loadpos commands in practice mode.
+- Fixed noclip not getting disabled after switching out from practice mode.
+- Fixed issues with rethrow of smoke / molly / nade.
+- Fixed .rethrow always throwing incendiary instead of molotov. 
+- Fixed `RemoteLogHeaderValue` value while resetting the match.
+
 # 0.8.13
 
 #### September 03, 2025